# wasi: proxy functions to that natural results are visible