--- a/imports/wasi_snapshot_preview1/args.go
+++ b/imports/wasi_snapshot_preview1/args.go
@@ -81,20 +81,17 @@ func argsGetFn(ctx context.Context, mod api.Module, params []uint64) Errno {
 // See argsGet
 // See https://github.com/WebAssembly/WASI/blob/snapshot-01/phases/snapshot/docs.md#args_sizes_get
 // See https://en.wikipedia.org/wiki/Null-terminated_string
-var argsSizesGet = newHostFunc(argsSizesGetName, argsSizesGetFn, []api.ValueType{i32, i32}, "result.argc", "result.argv_len")
+var argsSizesGet = proxyResultParams(&wasm.HostFunc{
+	Name:        "argsSizesGet",
+	ResultTypes: []api.ValueType{i32, i32, i32},
+	ResultNames: []string{"argc", "argv_len", "errno"},
+	Code:        &wasm.Code{IsHostFunction: true, GoFunc: u32u32ResultParam(argsSizesGetFn)},
+}, argsSizesGetName)
 
-func argsSizesGetFn(ctx context.Context, mod api.Module, params []uint64) Errno {
+func argsSizesGetFn(_ context.Context, mod api.Module, _ []uint64) (argc, argvLen uint32, errno Errno) {
 	sysCtx := mod.(*wasm.CallContext).Sys
-	mem := mod.Memory()
-	resultArgc, resultArgvLen := uint32(params[0]), uint32(params[1])
-
-	// argc and argv_len offsets are not necessarily sequential, so we have to
-	// write them independently.
-	if !mem.WriteUint32Le(ctx, resultArgc, uint32(len(sysCtx.Args()))) {
-		return ErrnoFault
-	}
-	if !mem.WriteUint32Le(ctx, resultArgvLen, sysCtx.ArgsSize()) {
-		return ErrnoFault
-	}
-	return ErrnoSuccess
+	argc = uint32(len(sysCtx.Args()))
+	argvLen = sysCtx.ArgsSize()
+	errno = ErrnoSuccess
+	return
 }