--- a/imports/wasi_snapshot_preview1/environ.go
+++ b/imports/wasi_snapshot_preview1/environ.go
@@ -84,20 +84,18 @@ func environGetFn(ctx context.Context, mod api.Module, params []uint64) Errno {
 // See environGet
 // https://github.com/WebAssembly/WASI/blob/snapshot-01/phases/snapshot/docs.md#environ_sizes_get
 // and https://en.wikipedia.org/wiki/Null-terminated_string
-var environSizesGet = newHostFunc(environSizesGetName, environSizesGetFn, []api.ValueType{i32, i32}, "result.environc", "result.environv_len")
+var environSizesGet = proxyResultParams(&wasm.HostFunc{
+	Name:        "environSizesGet",
+	ResultTypes: []api.ValueType{i32, i32, i32},
+	ResultNames: []string{"environc", "environv_len", "errno"},
+	Code:        &wasm.Code{IsHostFunction: true, GoFunc: u32u32ResultParam(environSizesGetFn)},
+}, environSizesGetName)
 
-func environSizesGetFn(ctx context.Context, mod api.Module, params []uint64) Errno {
+func environSizesGetFn(_ context.Context, mod api.Module, _ []uint64) (environc, environvLen uint32, errno Errno) {
 	sysCtx := mod.(*wasm.CallContext).Sys
-	mem := mod.Memory()
-	resultEnvironc, resultEnvironvLen := uint32(params[0]), uint32(params[1])
 
-	// environc and environv_len offsets are not necessarily sequential, so we
-	// have to write them independently.
-	if !mem.WriteUint32Le(ctx, resultEnvironc, uint32(len(sysCtx.Environ()))) {
-		return ErrnoFault
-	}
-	if !mem.WriteUint32Le(ctx, resultEnvironvLen, sysCtx.EnvironSize()) {
-		return ErrnoFault
-	}
-	return ErrnoSuccess
+	environc = uint32(len(sysCtx.Environ()))
+	environvLen = sysCtx.EnvironSize()
+	errno = ErrnoSuccess
+	return
 }