Right now, developing and troubleshooting wasi host functions are difficult as they are required to be WebAssembly 1.0 compatible. Specifically, they can only return a single result, and that's always used for errno. Other results are passed as out-parameters (memory offsets to write results), and their values are not very useful in practice.

This change proxies functions so that we can see their results while still keeping compatibility with wasm signatures. It uses the same approach we use in gojs (`GOOS=js GOARCH=wasm go build...`). What happens is we proxy in wasm, so that the go function implementing the ABI function has natural multi-value returned.

Specifically, we get logging like this, which is better as you can see the values returned:
```diff
-       ==> wasi_snapshot_preview1.args_sizes_get(result.argc=1,result.argv_len=6)
-       <== ESUCCESS
+       --> wasi_snapshot_preview1.args_sizes_get(result.argc=1,result.argv_len=6)
+               ==> wasi_snapshot_preview1.args_sizes_get.go()
+               <== (argc=2,argv_len=5,ESUCCESS)
+       <-- ESUCCESS
```