@codefromthecrypt 2022-12-06
will leave this draft as I refactor as many WASI functions as I can. In the process I imagine a lot of the code will generalize as opposed to re-defining N helpers manually.