@codefromthecrypt 2022-12-06
next step: convert the function that proxies the function to be generic. ex claim N result parameters