@codefromthecrypt 2022-12-07
made progress with a number of file-related calls. once I back port all the calls whose out params can fit in 8 bytes (i64), I'll go back and clean up. Particularly, I'll try to clean up logging so that results that are flags are formatted when in wasi mode, as well any obvious cruft cleanup opportunities in the main code. need another day.