@mathetake 2022-12-08
how's the perf diff look like before vs this?