@codefromthecrypt 2022-12-08
bench stat in progress!

```bash
$ git checkout main;go test -run='^$' -bench=. -count=5 > old.txt; git checkout wrap-wasi; go test -run='^$' -bench=. -count=5 >new.txt; benchstat old.txt new.txt
```