@codefromthecrypt 2022-12-08
perf looks same across 5 runs. verified with bench stat because ad-hoc looked the same

```
name                               old time/op    new time/op    delta
_ArgsEnviron/args_get-16              119ns ± 2%     119ns ± 2%    ~     (p=1.000 n=5+5)
_ArgsEnviron/args_sizes_get-16        106ns ±16%      97ns ± 2%    ~     (p=0.056 n=5+5)
_ArgsEnviron/environ_get-16           123ns ± 4%     118ns ± 2%  -4.21%  (p=0.016 n=5+5)
_ArgsEnviron/environ_sizes_get-16    98.7ns ± 3%    96.7ns ± 2%    ~     (p=0.310 n=5+5)

name                               old alloc/op   new alloc/op   delta
_ArgsEnviron/args_get-16              24.0B ± 0%     24.0B ± 0%    ~     (all equal)
_ArgsEnviron/args_sizes_get-16        24.0B ± 0%     24.0B ± 0%    ~     (all equal)
_ArgsEnviron/environ_get-16           24.0B ± 0%     24.0B ± 0%    ~     (all equal)
_ArgsEnviron/environ_sizes_get-16     24.0B ± 0%     24.0B ± 0%    ~     (all equal)

name                               old allocs/op  new allocs/op  delta
_ArgsEnviron/args_get-16               2.00 ± 0%      2.00 ± 0%    ~     (all equal)
_ArgsEnviron/args_sizes_get-16         2.00 ± 0%      2.00 ± 0%    ~     (all equal)
_ArgsEnviron/environ_get-16            2.00 ± 0%      2.00 ± 0%    ~     (all equal)
_ArgsEnviron/environ_sizes_get-16      2.00 ± 0%      2.00 ± 0%    ~     (all equal)
```