@codefromthecrypt 2022-12-15
I caused an init performance regression as I was only looking at call performance before/after. I will fix this possibly with a different design that doesn't use proxy functions. For example, passing the module memory to the logging interceptor instead. Before I thought about this approach but hesitated due to risking an interceptor corrupting memory. I think this risk is better than higher overhead all the time.

```
name                       old time/op    new time/op    delta
Allocation/Compile-16        38.3ms ± 0%    38.3ms ± 0%      ~     (p=0.556 n=4+5)
Allocation/Instantiate-16     845µs ± 0%    1666µs ± 1%   +97.29%  (p=0.008 n=5+5)
Allocation/Call-16           4.04µs ± 2%    4.00µs ± 2%      ~     (p=0.310 n=5+5)

name                       old alloc/op   new alloc/op   delta
Allocation/Compile-16        20.3MB ± 0%    20.3MB ± 0%      ~     (p=0.421 n=5+5)
Allocation/Instantiate-16     568kB ± 0%    1043kB ± 0%   +83.70%  (p=0.008 n=5+5)
Allocation/Call-16            48.0B ± 0%     48.0B ± 0%      ~     (all equal)

name                       old allocs/op  new allocs/op  delta
Allocation/Compile-16          432k ± 0%      432k ± 0%      ~     (p=1.000 n=5+5)
Allocation/Instantiate-16     6.74k ± 0%    16.59k ± 0%  +146.09%  (p=0.008 n=5+5)
Allocation/Call-16             5.00 ± 0%      5.00 ± 0%      ~     (all equal)
```